Prepare Perl URI module code for conversion to Lua
I've removed some Perl-specific cruft, expanded tabs, and added Vi modelines.
Removed the POD documentation as well, since I don't want it in the way
while I'm porting things, and I'll have to write my own anyway, probably
in a different format.

Also removed the 'URI::URL' compatibility module.  We're getting a fresh
start so we don't need to be compatible with anything else.  Deleted some
tests which exclusively used that deleted module, and switched some others
to use modern 'URI' instead.  Got rid of the compatibility methods in the
'URI::Heuristic' module as well, and switched the test to use the modern
ones.

Deleted the tests which relied on 'Storable' being available.  There isn't
a Lua equivalent as far as I know, and I don't think it's a big priority in
this case because usually a simple stringification of the URI will do.