Convert most of the Perl URI code to Lua
Quite a few of the test programs now run sucessfully.  The makefile I've
added runs all the ones which have now work.  The others rely on code which
hasn't been ported yet.

Remove the function 'uri_escape_utf8' from URI.Escape, because Lua doesn't
have support for character encodings.  The 'uri_unescape' function now only
takes a single argument, not a list (the only place it was called with a list
was in the test suite).

The URI._generic class no longer implements the special options
'ABS_ALLOW_RELATIVE_SCHEME' and 'ABS_REMOTE_LEADING_DOTS'.  I thought it
would be better to keep the number of publicly accessible knobs to a minimum.

The '_port' method in URI._server never returns an empty string or the number
zero, both of which could happen with bad URIs in the Perl version.  You get
nil instead in that case.

The URI.urn.oid method 'oid' accepts a list of values in Perl, but my version
takes a single array of values, or a single string.  It always returns an
array (but you can call the 'nss' method to get the string version).

The methods which accept and return lists in Perl now use tables, used as
arrays or hashes as appropriate.  This particularly applies to some of the
methods in URI._query and URI.sip.

The code in the URI.sip module has been restructured.  The Perl version
uses complex regular expressions to split the URIs into their component
parts.  Since Lua's patterns make that hard, I've added private functions
'_disassemble' and '_assemble' to take the URI appart, returning the
individual bits that are interesting, and then put it back together again
after one of the pieces has been changed.