Fix a bug in my version of the generic 'query' method
I've now got the 't/abs.t' tests working fully, with this bug fix.  The
Perl version is OK, but this is a different algorithm since I don't have
a full regex library to hand.

For the test program I've also split the RFC text which is used as a source
of test data into a separate file, since Lua doesn't have an equivalent
of Perl's 'DATA' file handle.