Port URI.urn.isbn module to Lua
I'm now able to do this having finally gotten Lua-ISBN into a state where
I can rely on it.

The URI.urn.isbn method 'isbn_as_ean' has been removed because it relies on
a method 'as_ean' in the ISBN module, and that doesn't exist in either my Lua
port or the original Perl version.

The URI.urn.isbn methods 'isbn_publisher_code' and 'isbn_country_code' have
been removed because you can just call the appropriate methods on the ISBN
object returned by the 'isbn' method (which returns a string in the Perl
version).  Also, the 'isbn_country_code' method called the 'country_code'
method on the ISBN object, which doesn't exist (presumably it was an old
name for the 'group_code' method).