Port the LDAP classes in the URI module from Perl to Lua
This all works, or at least it behaves about the same as the Perl version,
but I've had to bodge the names of a couple of public functions which used
to start with '_'.  Those won't get imported by my mixin thing, so I've
temporarily called them 'TODO_whatever'.  They'll get renamed to something
else in the future, but that would have happened anyway because I don't
want methods which look like they should be private cluttering up the
public API.