Port the URI.gopher module from Perl to Lua
There were no tests for this so I wrote some of my own, based on a quick
reading of RFC 4266.  I ran the same tests against the Perl code to make
sure I got the same behaviour.  I have changed one thing: my best guess
(not knowing much about Gopher) is that the 'selector' method should
default to the empty string rather than nil if there's nothing for it
in the path, so that's what I've done.

This code used the optional LIMIT argument to Perl's 'split' function, so
I've implemented that feature in my own 'URI._split' function and added
tests for it.