Avoid using 'module' function in URI library
All the modules in the URI library now create their tables explicitly in
a local variable, rather than having 'module' give them a new environment
to put everything in.  They return the module table so that callers can
name it however they want, so there's nothing that creates global variables.

Now that the library files are run in the global environment, I've been
able to take out all the indirection through a local '_G' variable to get
at global values.  When one file needs to use something from another, like
the functions in 'URI.Escape', they require it and store the table returned
in a local variable.

The tests in 't/pristine.t' now run successfully, confirming that there's
no corruption of the global environment.

I've added a string called '_MODULE_NAME' to each module's table, containing
the name used to require it, to help identifying the class of an object
when debugging.

In 'URI._generic' I no longer bother caching the 'URI._segment' table.
Instead I just require it every time it's needed, because that's fast.

In 'URI._ldap' I've made the '_normalize_dn' function local.  It's not
used from outside the module file and isn't documented.

I had to change one of the tests in 't/generic.t' due to a problem I don't
fully understand, which seemed to have something to do with stringifying
classes.  I've changed it to use 'assert' instead of 'assert_equal' and
just compared the class table with the metatable of an object directly.
I think that's as valid as the old test and still works.