Start new from-scratch generic URI implementation with different design
Renamed from 'URI' to 'uri'.  I've started this from scratch, throwing away
the generic URI code derived from Perl, so that I can take a different
design approach, and also base this on RFC 3986 rather than the older ones.

For now, this just does normalization of generic URIs (no scheme-specific
knowledge yet).

My new design carefully validates the input, and will reject anything bad.
The intention is that when (if) I add mutator methods, they will do the
same thing, ensuring that what you get when you stringify the object will
always be a valid URI.

I also now normalize (canonicalize) on input.  The intention is that the
URI object will always represent the maximally normalized URI.  Of course,
there's a limit to how much of that I can do in terms of what schemes are
implemented (none so far).

URI objects now store the component parts of the URI instead of a single
string.  I have to take the input apart anyway to validate each part, and
I'm not just accepting a string and storing it like the Perl module does,
I'm changing it to a preferred form for storage.

I extended the URI.Escape function 'uri_unescape' to allow a limited
range of characters to be unescaped (leaving others escaped), but I didn't
end up using that except in a TODO comment.  I am doing the same thing
when normalizing the percent encoding, but I do it specially so that I can
normalize the hex codes to uppercase.  Might be some refactoring room there.