Basic generic URI mutator methods; more tests for generic URIs
The basic accessor methods like 'scheme' are now mutators, but untested and
without any validation yet.  There are some more tests that a generic URI
is broken up into the right bits and the accessor can return them.

Use the accessor to piece together the URI when generating a string of it,
so that subclasses can override how particular parts are represented.