Move utility functions from old 'URI' module into new 'URI/_util.lua'
This is so that they can be got by all the modules for particular URI
schemes, without me having to pollute the namespace of the baseclass
with them.

Also updated tests in 't/_util.lua' to use the new module, and they once
again pass.

The only change I've made to this code compared to the same functions in
the current master version of 'URI' is to remove the underscores from the
start of their names.  They no longer need to look private because they
have their own table to sit in.