Reimplement HTTP URI support with new URI module design
I've now implemented the subclassing, and got the 'http' and 'https' scheme
modules basically working, although they don't do much yet, just the basics.

Based on my current interpretation of the spec, I'm disallowing userinfo
in HTTP URIs, and normalizing the path to "/" if it's empty.  I'm also
leaving the port out if it's the default port for a scheme.

To make the scheme-specific stuff happen, the constructor now attempts to
load a module for the scheme and sets that to be the new objects class if
the module exists (otherwise it's the generic 'uri' class).  It also calls
the 'init' method on the new object and returns the results from that so
that it can report validation errors just like the constructor.