Reimplement URN module and implement ISSN URNs
The new version of 'URI.urn' does all the validation and canonicalization
it can for generic URNs.  It also does the same kind of class-changing
trick that the base class constructor does, and knows how to call 'init'
again when the NID changes the class of an object.

I've implemented the 'ISSN' NID because it's a simple one to try first.
It didn't exist before in the Perl module, so this is an exciting new
feature (which nobody will ever need).