Port 'uri.urn.isbn' to new library design
This is also less reliant on the 'ISBN' module.  It will do some basic
validation and normalization without it, but it won't try to check the
checksum or put the hyphens in (so the normalized ISBN URI might just
have the digits).  You can get the digits without hyphens with the new
'isbn_digits' method, which works whetherh the 'ISBN' module is installed
or not.  The only thing that will cause an error if it's not installed
is the 'isbn' method, which can get and set the value using an ISBN object.