Fix the 'scheme' and 'nid' mutator methods
These methods can both cause a change in the URI which makes its current
class no longer appropriate, so it has to have its class changed and be
reinitialized to apply appropriate normalization and validation.  This can
cause the mutator method to throw an exception, in which case care has to
be taken to preserve its original state, otherwise the object would be
broken.

None of this is likely to be useful, but I wanted it to work right just
for completeness.  This seems to be the least surprising behaviour as far
as I can see.  The implementation isn't likely to be all that fast, but
I don't think that matters in this edge case.