Port 'uri.urn.oid' module to new library design
This now validates correctly.  It doesn't normalize the URI anymore than the
generic URN module does, because there's nothing else to be done.

I've replaced the Perl module's 'oid' method with 'oid_numbers' for
consistency with 'isbn_digits' and so on, and it now validates its input
to preserve objects in a correct state at all times.