Tighten up mutators for URN modules
The 'uri.urn.oid' and 'uri.urn.isbn' modules now have 'nss' methods to
validate and normalize the new value when it's set.

Slightly better error messages when initializing an invalid ISBN URI, or
when calling 'isbn_digits' with a bad value.

The 'path' mutator in 'uri.urn' now correctly changes class when necessary
and calls the NID-specific 'nss' method for validation and normalization.