Remove the non-standard and rare 'tn3270' and 'rlogin' URI schemes
The 'tn3270' scheme name is mentioned in RFC 1738 section 4, but as far
as I can discover has never been actually defined.  I can find no mention
of an 'rlogin' scheme in any RFC or draft.

I've never heard of either of these schemes being actually used, so I
don't think we need them.  It looks like the Perl URI module has them just
as an extrapolation from the 'telnet' one, which is defined.  The Perl
implementation's changelog says that 'tn3270' was added in 2003, and 'rlogin'
in 1998, and there's no mention of them since.