Fix bugs in 'uri' for port normalization and userinfo mutator
The normalization of an explicit port to an implicit one (where the scheme
class has a 'default_port' method) was broken, since I was checking for a
default port before the scheme-specific class was set.

The 'userinfo' mutator now does percent-encoding normalization and checks
for the presence of extra '@' characters, which aren't allowed (they would
make the URI invalid when it was serialized unless I specially escaped them,
but since the caller is expected to do other percent encoding for userinfo
it seems to make sense to require this as well).