Provide support for 'username:password' in URIs; fix 'uri.telnet' module
Support for the 'telnet' scheme now works with the new library design.
It does normalization and validation properly.

In order to make this work I had to provide support for dividing a userinfo
value in a URI into separate username and password fields, separated by a
colon.  This is used in various URI schemes, but currently telnet is the
only one which uses this support properly in this library.

I've made it illegal to have more than one ':' in a URI which uses the
'username:password' convention.  This seems to be implied by the older
RFCs, although not mentioned in the new ones.  The alternative would have
been to make an arbitrary decision about whether to choose the first or
last colon as the delimiter.

Scheme modules which support username/password pairs should subclass the
'uri._login' module, just as in the original Perl design, but I've simplified
things by moving the 'username' and 'password' methods into that class
rather than adding them in from the mixin class 'uri._userpass' (which can
now be removed).  The 'username' method was called 'user' in the Perl
version.

I've added tests for 'telnet' URIs, and for the username/password support
in general, all in 'test/telnet.lua'.

I've removed the 'uri._server' class, which 'uri._login' used to derive
from.  Now it can subclass 'uri' directly, because the old Perl '_server'
class only contains methods that are provided in the base class.

The new version of '_login' has a bit of a kludge, in the form of a new
method 'init_base', which is called by the 'init' method in the 'telnet'
class.  This should probably be part of a more general convention in the
class hierarchy, but I can't see how to arrange it yet.