Move everything from 'uri.Escape' into 'uri._util'
This helps to move towards a consistent naming convention where all the
module names are in lowercase (as is usual for Lua) and everything under
'uri' which isn't for a particular URI scheme has a name starting with '_'.

There were only two functions and a private table of escapes sequences in
the file 'uri/Escape.lua' anyway, so I might as well consolidate it with
the more miscelaneous utility stuff.