Port 'uri.data' module to new library design
This now does some validation, although it doesn't check the syntax of the
media type and its parameters yet.  It doesn't do any normalization yet.

I've renamed the 'data' method to 'data_bytes', and 'media_type' to
'data_media_type', both to avoid cluttering the method namespace with
very specific things.  I'm following the same basic convention in other
classes.

Unlike the old Perl version, we return a media type starting with the
string 'text/plain' if the one in the URI starts with a semicolon (which
is an abbreviation like ';charset=...'.

The choice of which characters get escaped and unescaped is made based on
the character sets defined in RFC 3986 rather than the older ones.

I provide an overridden 'path' method to validate the path for this scheme,
and I override 'userinfo', 'host', and 'port' to prevent those being set,
since they're not allowed for 'data' URIs.

This should now work better when my 'datafilter' module isn't installed.
Most of the tests should run without it now.  The only thing it is needed
for is to decode a URI which has got base64 data in.  URI encoding will be
used for encoding if it's not available.