Remove Perl URI modules which aren't going to get ported to Lua yet
The URI::Heuristic and URI::WithBase modules are extra features which I
don't think are very important, and certainly aren't needed for the rest
of the library to be useful, so they can go for now at least.

The features in the URI::QueryParam module might get provided in due course,
but if so they'll be integrated elsewhere, not in their own module.