Allow relative URIs to be resolved against a base URI
There is now a 'resolve' method on all URI classes.  It does nothing except
in the 'uri._relative' class, where it adjusts the URI in place to make it
absolute.

The 'new' method now takes an optional base URI argument which uses 'resolve'
to resolve the main URI.

The tests in 'test/abs.lua' have been adjusted to use these two new APIs,
and the test data in 'test/data/abs-data.txt' has been updated to the text
from RFC 3986.  The old test data from 'draft-fielding-uri-syntax-02' does
not work because the expected results have changed for a few tests.

The code for implementing the 'remove_dot_segments' algorithm has been
moved from the 'uri' class into a function in 'uri._util'. so that it can
be used in two places.