Remove unnecessary 'roy-test' tests
The test data in 'test/data/roytest1.html' is the same as that tested by
'test/_resolve.lua' already, and the others may not be up to date with the
current specs, so I'm removing them all for now.

Also removed a redundant item in 'test/data/abs-data.lua' and tweaked the
whitespace in there.