Remove 'clone' method, just use constructor instead
The 'clone' method is unnecessary when calling 'new' with a URI object does
the same thing, and I want to keep the API as small as possible to start with.

Updated subclasses and tests to match.  While I'm at it, renamed the file
'test/generic.lua' to 'test/_generic.lua' to be consistent with the scheme
of using underscore prefixes for test programs which don't test a specific
URI scheme.  Removed the 'test/clone.lua' file because it duplicates the
generic tests.