Add support for FTP URIs
I've based this on RFC 1738 section 3.2, as there doesn't seem to be a more
up to date reference for the FTP scheme.  It does some very simple
normalization, and provides easy access to the 'type' parameter with the
'ftp_typecode' method, but I haven't tried to validate the parameter since
the RFC suggests other FTP features could be supported by URIs in undefined
ways.