Provide 'relativize' method on URI objects
New method to change a URI object to contain a relative URI reference instead
of an absolute one, similar to the 'rel' method in the Perl URI library,
except it does its changes in place rather than creating a new object.

This was originally going to be called 'unresolve', and I considered
'make_relative' for a while.