Move test data into test program file for simplicity
I thought I was going to use this data in more than one test program, and
it was originally a text file, but now that it's Lua code only used in one
place I might as well fold it back into the test code.  This leaves me with
no extra test data files, which is nice.