Fix URI resolving when the resolved URI would be invalid
The 'resolve' method correctly threw an exception when this happened, but it
left the object with the wrong metatable, and potentially other wrongness.
It now stores a backup of the original URI as a string, and then restores
the object to its original state if something goes wrong, before allowing
the exception to reach the caller.

When the constructor is used to resolve a relative URI it was also throwing
an exception, but for consistency it now returns nil and an error message.