Fully implement 'host', 'query', and 'fragment' methods for URIs
Instead of basing these on a generic function they are now fully implemented
to do proper error checking and normalization.  To do checking of host
syntax I've factored out the code for that into a new internal function
'_is_valid_host'.

Also, the 'port' and 'userinfo' methods now create an empty host if there
is none, otherwise the information cannot be presented properly in the URI
when it is serialized.