Refactor and correct disallowing certain URI parts for certain schemes
I've factored out the code which created methods in the 'data', 'file', and
'urn' scheme classes to prevent callers from setting parts of the URI which
should never be present (for example a query part for URNs, or a port for
a file URI).  These classes now make calls to the function
'uri_part_not_allowed' in the 'uri._util' module.

The 'uri.http' module now uses this mechanism to disallow 'userinfo' parts,
as it should have been doing already.

I've corrected the method which gets installed to override the generic
accessor methods.  It used to always return nil, but that turns out to
be wrong (the new test in 'test/http.lua' detects the problem), because
during the 'init' method when changing the scheme from a class which allows
a part to one which doesn't, the overridden method hides the fact that there
is a part in the URI which shouldn't be there, so the scheme-specific
validation can't detect it.