Slightly better error messages from within methods
I've added appropriate stack level values to most calls to the 'error'
function, when calling it to report bad usage, so that the error points
to the place where the user's code is calling into my libraries.  There
are a few places where this level won't be enough (such as when a method
gets called from within a subclass overridden version of the same thing,
which adds an extra level of stack thing), but it will usually be right.