Add copies of Lunit testing code
These files are distributed in my packages, and are required to run the
tests, so it makes life easier to commit them here, even though I don't
maintain them.

All the 'lunit.lua' files are from Lunit 0.3alpha, and are unchanged,
except for the one in 'datetime', which is from Lunit 0.4, and has a
minor change to fix a bug on line 573.  The 0.4 version also requires
the other two files I'm committing in order to be run.